+++ pom.xml
@@ -11,7 +11,7 @@
     </parent>
     <groupId>com.citrsw</groupId>
     <artifactId>java-api-docs</artifactId>
-    <version>1.5.4-beta</version>
+    <version>1.5.5-beta</version>
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
