+++ src/main/java/com/citrsw/common/ApiUtils.java
@@ -1,6 +1,13 @@
 package com.citrsw.common;
 
+import sun.reflect.generics.reflectiveObjects.ParameterizedTypeImpl;
+
+import java.lang.reflect.ParameterizedType;
+import java.lang.reflect.Type;
+import java.net.InetAddress;
+import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 /**
@@ -19,7 +26,7 @@ public class ApiUtils {
         TYPE_MAP.put("Long", "Int");
         TYPE_MAP.put("long", "Int");
         TYPE_MAP.put("Boolean", "Bool");
-        TYPE_MAP.put("boolean", " Bool ");
+        TYPE_MAP.put("boolean", "Bool");
         TYPE_MAP.put("Float", "Float");
         TYPE_MAP.put("float", "Float");
         TYPE_MAP.put("Double", "Double");
@@ -34,6 +41,7 @@ public class ApiUtils {
         TYPE_MAP.put("MultipartFile", "[Int]");
     }
 
+
     /**
      * java类型转Ios类型
      *
@@ -42,6 +50,35 @@ public class ApiUtils {
      */
     public static String javaToIosType(String className) {
         return TYPE_MAP.get(className);
+    }
+
+    /**
+     * 重新生成符合泛型形式的Type
+     * 用于返回形式的泛型
+     *
+     * @param returnType 返回的Type
+     * @return 符合泛型形式的Type
+     */
+    public static Type regenerateType(Type returnType) {
+        if (returnType instanceof ParameterizedType) {
+            // 强制类型转换
+            ParameterizedType pType = (ParameterizedType) returnType;
+            Type rawType = pType.getRawType();
+            List<Type> types = new ArrayList<>();
+            for (Type actualTypeArgument : pType.getActualTypeArguments()) {
+                types.add(regenerateType(actualTypeArgument));
+            }
+            return ParameterizedTypeImpl.make((Class<?>) rawType, types.toArray(new Type[0]), null);
+        } else {
+            return returnType;
+        }
+    }
+    public static String getLocalIp() throws Exception{
+        String localIp = "";
+        InetAddress addr = InetAddress.getLocalHost();
+        //获取本机IP
+        localIp = addr.getHostAddress();
+        return localIp;
     }
 
 }
